% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{getdata}
\alias{getdata}
\title{Load Sample Data}
\usage{
getdata(...)
}
\arguments{
\item{...}{String name of dataset}
}
\value{
Dataframe
}
\description{
This function loads data from a package as a dataframe. It assumes that the 
package is already loaded with library() and expects a string that is the
name of the dataset.
}
