% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{new_NASApredictions}
\alias{new_NASApredictions}
\title{Make New Predictions}
\usage{
new_NASApredictions(specified_model, new_data)
}
\arguments{
\item{specified_model}{String name of acceptable model type listed in details}

\item{new_data}{Dataframe of new data with all sensor measurements}
}
\value{
Predictions
}
\description{
This function makes predictions on new data. For every rown in the dataset, a prediction will 
be made. There are 8 models that have been trained and are ready to make regression and classification
predictions. For regression the predictions will be the RUL (the remaining useful life of the engine)
and for classification the predictions will be whether or not the engine will fail in the next 30
days.
}
\details{
The models that the user can choose from are: basic_classtree_model, basic_linear_model,                
basic_logistic_model, basic_regtree_model, tuned_rf_class_model, tuned_rf_reg_model, tuned_xgb_class_model,              
tuned_xgb_reg_model.
}
